+++ code/src/vs/workbench/browser/parts/titlebar/commandCenterControl.ts
@@ -181,7 +181,10 @@ class CommandCenterCenterViewItem extends BaseActionViewItem {
 						private _getLabel(): string {
 							const { prefix, suffix } = that._windowTitle.getTitleDecorations();
 							let label = that._windowTitle.workspaceName;
-							if (that._windowTitle.isCustomTitleFormat()) {
+							const header = that._windowTitle.getHeader();
+							if (header) {
+								label = header;
+							} else if (that._windowTitle.isCustomTitleFormat()) {
 								label = that._windowTitle.getWindowTitle();
 							} else if (that._editorGroupService.partOptions.showTabs === 'none') {
 								label = that._windowTitle.fileName ?? label;